requirements.txt
aniso8601==8.0.0
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pgdb==0.0.8
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.4.2
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
